started main project work
started printing variables into the file need to fetch out all fields that are database fields and fill them into a struct to then have all meta data required to generate functions. 

also need to workout how to do embedding of structs so that i can ref that, and allow a different api face compared to the database object. and they auto combine.