Generator funcs Insert,Update & delete Complete
all three funcstions and their multi variant are complete but not yet tested. 

also constants in an attempt to setup the get command is not yet working.
working out how to allow easy wildcard searching straight from a proto service call into an api call to allow easy dynamic searching is proving to be difficult to workout.